"Refactor version management imports"
