+++ cli/installations.ts
@@ -2,14 +2,14 @@ import {
   APP_VERSION,
   IdentifierKind,
   type XmtpEnv,
-} from "version-management/client-versions";
+} from "version-management/node-versions";
 import "dotenv/config";
 import {
   createSigner,
   generateEncryptionKeyHex,
   getEncryptionKeyFromHex,
 } from "@helpers/client";
-import { Client } from "version-management/client-versions";
+import { Client } from "version-management/node-versions";
 import { generatePrivateKey } from "viem/accounts";
 
 interface Config {
