+++ helpers/streams.ts
@@ -8,7 +8,7 @@ import {
   type Conversation,
   type Dm,
   type Group,
-} from "version-management/client-versions";
+} from "version-management/node-versions";
 
 // Define the expected return type of verifyMessageStream
 export type VerifyStreamResult = {
