+++ inboxes/bysize.test.ts
@@ -1,6 +1,6 @@
 import { ProgressBar } from "@helpers/logger";
 import { getWorkers, type Worker } from "@workers/manager";
-import { IdentifierKind } from "version-management/client-versions";
+import { IdentifierKind } from "version-management/node-versions";
 import { describe, it } from "vitest";
 
 describe("bysize", () => {
