+++ monitoring/bugs/stitch.test.ts
@@ -1,6 +1,6 @@
 import { getInboxes } from "@inboxes/utils";
 import { getWorkers } from "@workers/manager";
-import { type Group } from "version-management/client-versions";
+import { type Group } from "version-management/node-versions";
 import { describe, expect, it } from "vitest";
 
 const testName = "sync";
