+++ monitoring/forks/forks.test.ts
@@ -2,10 +2,7 @@ import { getTime } from "@helpers/logger";
 import { setupDurationTracking } from "@helpers/vitest";
 import { getInboxes } from "@inboxes/utils";
 import { getWorkers, type Worker } from "@workers/manager";
-import {
-  getActiveVersion,
-  type Group,
-} from "version-management/client-versions";
+import { getActiveVersion, type Group } from "version-management/node-versions";
 import { describe, it } from "vitest";
 
 // Count of groups to create
