+++ monitoring/networkchaos/group-client-partition.test.ts
@@ -1,7 +1,7 @@
 import { verifyMessageStream } from "@helpers/streams";
 import { setupDurationTracking } from "@helpers/vitest";
 import { getWorkers } from "@workers/manager";
-import type { Group } from "version-management/client-versions";
+import type { Group } from "version-management/node-versions";
 import { describe, expect, it } from "vitest";
 import * as iptables from "../../network-stability/iptables";
