+++ monitoring/performance.test.ts
@@ -19,7 +19,7 @@ import {
   IdentifierKind,
   type Dm,
   type Group,
-} from "version-management/client-versions";
+} from "version-management/node-versions";
 import { describe, expect, it } from "vitest";
 
 const testName = "performance";
